requirements.txt
Django==4.2.3
djangorestframework==3.14.0
channels==4.0.0
daphne==4.0.0
psycopg2-binary==2.9.6
Pillow==9.5.0
django-cors-headers==4.1.0
python-dotenv==1.0.0